ref: Refactor event hashing to be more reusable
