+++ tests/sentry/event_manager/test_get_hashes_from_fingerprint.py
@@ -4,8 +4,8 @@ from __future__ import absolute_import, print_function
 
 import mock
 
-from sentry.event_manager import get_hashes_for_event
 from sentry.models import Event
+from sentry.event_hashing import md5_from_hash
 
 
 @mock.patch('sentry.interfaces.stacktrace.Stacktrace.compute_hashes')
@@ -29,9 +29,9 @@ def test_stacktrace_wins_over_http(http_comp_hash, stack_comp_hash):
         platform='python',
         message='Foo bar',
     )
-    hashes = get_hashes_for_event(event)
+    hashes = event.get_hashes()
     assert len(hashes) == 1
     hash_one = hashes[0]
     stack_comp_hash.assert_called_once_with('python')
     assert not http_comp_hash.called
-    assert hash_one == ['foo', 'bar']
+    assert hash_one == md5_from_hash(['foo', 'bar'])
